requirements.txt
av
absl-py==0.12.0
astunparse==1.6.3
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cloudpickle==1.6.0
cycler==0.10.0
Cython==0.29.22
decorator==5.0.6
easydict==1.9
Flask==1.1.2
flatbuffers==1.12
gast==0.4.0
gevent==21.1.2
google-auth==1.28.1
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
greenlet==1.0.0
grpcio==1.37.0
h5py==3.2.1
hrnet-pose-estimation@git+https://github.com/ovs-code/HRNet-Human-Pose-Estimation.git@ddba0bd95a96bc9e95183af5ad172cb7c1fb24e8
idna==3.1
imageio==2.9.0
itsdangerous==1.1.0
Jinja2==2.11.3
json-tricks==3.15.5
jsonpickle==2.0.0
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.4.1
mlq==0.2.2
msgpack==0.5.6
networkx==2.5.1
numpy==1.20.2
oauthlib==3.1.0
opencv-python==4.5.1.48
opt-einsum==3.3.0
pandas==1.2.3
Pillow==8.2.0
protobuf==3.15.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
pytzdata==2020.1
PyWavelets==1.1.1
PyYAML==5.4.1
redis==2.10.6
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.18.1
scipy==1.6.2
Shapely==1.7.1
six==1.15.0
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow-estimator==2.4.0
tensorflow-gpu==2.4.1
termcolor==1.1.0
tifffile==2021.4.8
torch==1.8.1
torchvision==0.9.1
tqdm==4.60.0
typing-extensions==3.7.4.3
urllib3==1.26.4
Werkzeug==1.0.1
wrapt==1.12.1
yacs==0.1.8
zope.event==4.5.0
zope.interface==5.3.0
numba
